requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.64.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# For model loading and inference
accelerate>=0.20.0
bitsandbytes>=0.37.0
safetensors>=0.3.0
tokenizers>=0.13.0

# For evaluation and metrics
datasets>=2.0.0
rouge-score>=0.1.0
bert-score>=0.3.13

# For configuration and utilities
pyyaml>=6.0
jinja2>=3.1.0
jsonlines>=3.0.0
fire>=0.5.0
omegaconf>=2.3.0

# For API integrations (optional)
openai>=1.0.0
anthropic>=0.25.0
together>=0.2.0
ollama>=0.1.7
litellm>=1.0.0
dspy-ai>=2.4.0

# Text processing
nltk>=3.7
spacy>=3.4.0
beautifulsoup4>=4.11.0
requests>=2.28.0

# Async and concurrency
aiohttp>=3.8.0

# Memory and performance
vllm>=0.2.0  # Optional: for efficient inference
fastapi>=0.95.0  # For API serving
uvicorn>=0.20.0  # For API serving

# For visualization and reporting
plotly>=5.0.0
kaleido>=0.2.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.0.0

# System utilities
psutil>=5.9.0
GPUtil>=1.4.0